move test directory
