+++ src/components/admin/schulen/OrganisationUnassign.spec.ts
@@ -4,8 +4,8 @@ import type { Organisation } from '@/stores/OrganisationStore';
 import type { Person } from '@/stores/types/Person';
 import { PersonWithZuordnungen } from '@/stores/types/PersonWithZuordnungen';
 import type { Zuordnung } from '@/stores/types/Zuordnung';
-import { DoFactory } from '@/testing/DoFactory';
 import { flushPromises, mount, VueWrapper } from '@vue/test-utils';
+import { DoFactory } from 'test/DoFactory';
 import type { MockInstance } from 'vitest';
 import { nextTick } from 'vue';
 import { createRouter, createWebHistory, type Router } from 'vue-router';
