+++ src/stores/OrganisationStore.spec.ts
@@ -1,9 +1,9 @@
 import { OrganisationsTyp, type OrganisationRootChildrenResponse } from '@/api-client/generated';
 import ApiService from '@/services/ApiService';
-import { DoFactory } from '@/testing/DoFactory';
 import { rejects } from 'assert';
 import MockAdapter from 'axios-mock-adapter';
 import { createPinia, setActivePinia } from 'pinia';
+import { DoFactory } from 'test/DoFactory';
 import {
   SchuleType,
   useOrganisationStore,
