+++ tsconfig.app.json
@@ -8,7 +8,7 @@
     "module": "ES2020",
     "moduleResolution": "Node",
     "paths": {
-      "@/*": ["./src/*"],
+      "@/*": ["./src/*"]
     },
     "resolveJsonModule": true,
     "types": ["vitest/globals"],
