+++ vite.config.ts
@@ -42,6 +42,7 @@ export default defineConfig({
   resolve: {
     alias: {
       '@': fileURLToPath(new URL('./src', import.meta.url)),
+      test: fileURLToPath(new URL('./test', import.meta.url)),
     },
   },
   server: {
